backend/requirements.txt
fastapi
uvicorn
coqui-tts
pexels-api
ffmpeg-python
python-dotenv
requests
